+++ contracts/account-history-contract/src/action/mod.rs
@@ -1,6 +1,6 @@
 pub mod sudo {
-    mod update_account;
-    pub use update_account::{
+    mod update_account_chain;
+    pub use update_account_chain::{
         clean_old_history, clean_up_history, update_account, update_account_chain,
     };
 }
